crm: fix the four gates rung 7 was landing red on
Every failing job on this branch traced to one of four causes, each the new
rung diverging from a convention the ladder already had:

- `-Wuseless-cast` (GCC) in the extension-bag spike: `size()` already
  returns the type it was cast to. Broke every GCC leg plus Valgrind and
  the ladder build.
- `-Wswitch-default` (Clang `-Weverything`) in `crm::roleToString`: the
  switch is exhaustive and fell through to a trailing return, but the
  warning wants the arm inside. kanban::roleToString has the identical
  function with the arm and the comment saying why; crm now matches.
  Broke clang-coverage, the ladder sanitizer leg, and all-optional-features.
- The journal-stamp gate: `SelfJournal::recordRejectedPayload` is correctly
  unstamped and said so in its doc block, but the checker wants a
  `journal-stamp-exempt:` marker on the body, where lims's identical
  overload carries it. Moved the reasoning there.
- The rung-filter gate: `crm` reached examples/rungs.txt but not the two
  files that cannot derive from it -- wasm-ladder.yml's two literal path
  lists and codecov.yml's components.

The codecov component carries a measured figure on the same terms as
ledger's and lims's: 1468/1616 lines, 90.84%, target 89% and informational.
Its comment also records what the measurement turned up and does not fix:
ContactModel::attachActionLog and SavedViewModel::attachActionLog are
called by no test, so those two models' journaling paths are unverified
while the other four each have an explicit journaling test.

Verified locally in the configuration that failed (clang-coverage,
-Weverything, MORPH_BUILD_LADDER=ON): build clean, 2354/2354 tests pass
(crm's own 176 among them), and all six drift-guard gates plus their
self-tests pass.

+++ tests/test_extension_bag_spike.cpp
@@ -233,7 +233,12 @@ BRIDGE_REGISTER_ACTION(EB_ContactModel, EB_UpdateContact, "EB_UpdateContact")
 
     std::uint64_t nextOrder = 0;
     if (dom.contains("properties") && dom["properties"].holds<glz::generic_u64::object_t>()) {
-        nextOrder = static_cast<std::uint64_t>(dom["properties"].get<glz::generic_u64::object_t>().size());
+        // No cast: `size()` returns `std::size_t`, which is the same type as
+        // `std::uint64_t` on the 64-bit targets this file is built for, so an
+        // explicit cast is a no-op GCC rejects under -Wuseless-cast. Were
+        // `std::size_t` ever narrower, the assignment widens implicitly and
+        // safely -- unsigned to wider unsigned preserves the value.
+        nextOrder = dom["properties"].get<glz::generic_u64::object_t>().size();
     }
 
     for (auto const& field : fields) {
